Se maneja con paneles la interfaz, la que se ve azul es para que la reemplace por el pintados de los pispirispis. Ya se escoge cualquier grafica y se pinta.

+++ src/logic/Data.java
@@ -2,7 +2,7 @@ package logic;
 
 public class Data {
 	
-	private double day; 
+	private int day; 
 	private double value; 
 	
 	public Data(int day, double value) {
@@ -11,10 +11,10 @@ public class Data {
 		this.value = value;
 	}
 	
-	public double getDay() {
+	public int getDay() {
 		return day;
 	}
-	public void setDay(double day) {
+	public void setDay(int day) {
 		this.day = day;
 	}
 	public double getValue() {
